+++ Back_end/src/main/Project/parseEvaluator/nodes/BlockStatementNode.java
@@ -13,7 +13,7 @@ public class BlockStatementNode implements Node{
     }
 
     @Override
-    public double evaluate(Map<String, Integer> bindings){
+    public long evaluate(Map<String, Integer> bindings){
         for(Node node : statements){
             node.evaluate(bindings);
         }
