+++ Back_end/src/main/Project/parseEvaluator/nodes/Node.java
@@ -2,6 +2,8 @@ package Project.parseEvaluator.nodes;
 import java.util.Map;
 
 public interface Node {
-    double evaluate(Map<String,Integer> bindings);
+    long evaluate(Map<String,Integer> bindings);
     void print(int height, Map<String,Integer> bindings);
+
+
 }
